refactor(server): split _server_tools.py into a facet-aligned _server_tools/ package
register_tools() decomposed into reader/graph/index/writer/git submodules,
each exposing register(mcp), orchestrated by the package __init__. Shared
read-side helpers in _common.py. Pure code-move: tool bodies verbatim, no
manifest/behavior change (pinned by TestToolManifest). server.py untouched.

The file rename pulled two mechanical follow-ons: the ruff per-file-ignore
entry repoints _server_tools.py -> _server_tools/*.py, and a test that
imports _reindex_after_pull repoints to _server_tools.git.

Closes #578
