+++ pyproject.toml
@@ -120,7 +120,7 @@ ignore = ["E501"]
 "src/markdown_vault_mcp/cli.py" = ["B008"]
 "src/markdown_vault_mcp/server.py" = ["B008", "TC002"]
 "src/markdown_vault_mcp/_server_deps.py" = ["B008", "TC002", "TC003"]
-"src/markdown_vault_mcp/_server_tools.py" = ["B008", "TC001", "TC002"]
+"src/markdown_vault_mcp/_server_tools/*.py" = ["B008", "TC001", "TC002"]
 "src/markdown_vault_mcp/_server_resources.py" = ["B008", "TC001", "TC002"]
 "src/markdown_vault_mcp/_server_prompts.py" = ["B008", "TC002"]
 "src/markdown_vault_mcp/_server_apps.py" = ["B008", "TC001", "TC002"]
